Moved AllTeamsView to views/team folder
